[MOLECULE] FIX: Use the correct base images for Alpine Linux.
This commit fixes the fact that the `apache2-alpine-3.11` and
`apache2-alpine-3.10` images where using the `alpine:3.9` image
as a base instead of the correct ones.

This change also required some changes in the playbook and the
helper xsendfile.py script due to the usage of python3 by the
apache2 application on 'alpine:3.11'.